+++ packages/cli/__tests__/zoslogs/__system__/list/__scripts__/logs/list_logs_successful_rfj.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+zowe zos-logs list logs --rfj
+exit $?