+++ packages/cli/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
-# Recent Changes
+## Recent Changes
+
+- Enhancement: Added new feature to manage zos-logs. z/OSMF version 2.4 or higher is required. Ensure that the [z/OSMF Operations Log Support is available via APAR and associated PTFs](https://www.ibm.com/support/pages/apar/PH35930). [Issue 1104](https://github.com/zowe/zowe-cli/issues/1104)
+
+## `6.36.1`
 
 - BugFix: Fixed an issue where plugin install and uninstall did not work with NPM version 8. [#683](https://github.com/zowe/imperative/issues/683)